+++ src/main/java/pages/DemoWebshop_JewelsPage.java
@@ -9,13 +9,13 @@ public class DemoWebshop_JewelsPage {
     public static By txtbx_LngthInsideFrstItm = By.xpath("(//input[@type='text'])[3]");
     public static By btn_AddCartInsideFrstItm=By.xpath("(//input[@value='Add to cart'])[1]");
 
-    public static void adding_firstJewelToCart(String jewel_Length){
+    public static void adding_firstJewelToCart(String jewel_Length)throws Exception{
         WebDriver driver = BaseClass.getDriver();
         driver.findElement(DemoWebshop_JewelsPage.link_FirstItem).click();
         System.out.println("Clicked on first item in jewellery ");
 
         driver.findElement(DemoWebshop_JewelsPage.txtbx_LngthInsideFrstItm).sendKeys(jewel_Length);
-
+        Thread.sleep(2000);
         driver.findElement(DemoWebshop_JewelsPage.btn_AddCartInsideFrstItm).click();
         System.out.println("Clicked on Add to Cart Button in Jewellery");
     }
