+++ src/main/java/pages/OrangeHRM_EmployeeManagementPage.java
@@ -2,7 +2,10 @@ package pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import utilities.BaseClass;
+
+import java.util.List;
 
 public class OrangeHRM_EmployeeManagementPage {
     public static By Link_FirstEmployee = By.xpath("//tr[contains(@ng-repeat,'employee.empNumber')][1]/td[2]");
@@ -19,5 +22,21 @@ public class OrangeHRM_EmployeeManagementPage {
 
         driver.findElement(OrangeHRM_EmployeeManagementPage.Link_PersonalDetails).click();
         System.out.println("Clicked on Personal Details tab");
+    }
+
+    public static String finding_supervisorName(String employee_name){
+
+        WebDriver driver = BaseClass.getDriver();
+        String supervisor_name = null;
+        List<WebElement> employee_List = driver.findElements(OrangeHRM_EmployeeManagementPage.list_employee);
+        for (WebElement employee_Ref : employee_List) {
+            String employee = employee_Ref.getText();
+            if (employee.equalsIgnoreCase(employee_name)) {
+                supervisor_name = employee_Ref.findElement(OrangeHRM_EmployeeManagementPage.txt_supervisorName).getText();
+                System.out.println("Supervisor name of the given employee is :"+supervisor_name);
+                break;
+            }
+        }
+        return supervisor_name;
     }
 }
